+++ src/CMakeLists.txt
@@ -3,17 +3,21 @@ cmake_minimum_required(VERSION 2.8)
 file(GLOB dotconf_sources "../contrib/dotconf/src/*.c")
 
 file(GLOB sources "*.cpp")
-file(GLOB headers "*.h" "dba/*.h")
+file(GLOB headers "*.h" "db/database.h")
 
 # Conditionally add database sources
-if(${DBA_POSTGRES})
-  file(GLOB postgres_sources "dba/postgres.cpp")
+if(${DB_POSTGRES})
+  file(GLOB postgres_sources "db/postgres.cpp")
+  file(GLOB postgres_headers "db/postgres.h")
   set(sources ${sources} ${postgres_sources})
+  set(sources ${sources} ${postgres_headers})
 endif()
 
-if(${DBA_MONGO})
-  file(GLOB mongo_sources "dba/mongo.cpp")
+if(${DB_MONGO})
+  file(GLOB mongo_sources "db/mongo.cpp")
+  file(GLOB mongo_headers "db/mongo.h")
   set(sources ${sources} ${mongo_sources})
+  set(headers ${sources} ${mongo_headers})
 endif()
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} --std=c++11")
