+++ src/dazeus.cpp
@@ -4,7 +4,8 @@
  */
 
 #include "dazeus.h"
-#include "dba/database.h"
+#include "db/database.h"
+#include "db/factory.h"
 #include "network.h"
 #include "config.h"
 #include "plugincomm.h"
@@ -12,6 +13,10 @@
 #include <cassert>
 #include <sstream>
 #include <iostream>
+
+using dazeus::db::DatabaseConfig;
+using dazeus::db::Database;
+using dazeus::db::Factory;
 
 // #define DEBUG
 #define VERBOSE
@@ -103,40 +108,29 @@ bool dazeus::DaZeus::configLoaded() const
  */
 bool dazeus::DaZeus::connectDatabase()
 {
-    assert(config_->isRead());
-    const DatabaseConfig &dbc = config_->getDatabaseConfig();
+  assert(config_->isRead());
+  const DatabaseConfig &dbc = config_->getDatabaseConfig();
 
-#ifdef DBA_POSTGRES
-	if (dbc.type == "postgres" || dbc.type == "postgresql" || dbc.type == "pq" || dbc.type == "psql")
-	{
-		database_ = new PostgreSQLDatabase(dbc);
-	}
-#endif
-#ifdef DBA_MONGO
-	if (dbc.type == "mongo" || dbc.type == "mongodb") {
-		database_ = new MongoDatabase(dbc);
-	}
-#endif
+  try {
+    database_ = Factory::createDb(dbc);
+  } catch (dazeus::db::exception &e) {
+    std::cout << e.what() << std::endl;
+    return false;
+  }
 
-	if (!database_) {
-	  fprintf(stderr, "Unknown database type: %s\n", dbc.type.c_str());
-	  return false;
-	} else {
-	  try {
-        database_->open();
-      } catch(Database::exception &e) {
-        fprintf(stderr, "Could not connect to database: %s\n", e.what());
-        return false;
-      }
-	}
-
+  try {
+    database_->open();
+  } catch(dazeus::db::exception &e) {
+    std::cout << "Could not connect to database: " << e.what() << std::endl;
+    return false;
+  }
 	return true;
 }
 
 /**
  * @brief Return the database.
  */
-dazeus::Database *dazeus::DaZeus::database() const
+Database *dazeus::DaZeus::database() const
 {
   return database_;
 }
