+++ src/plugincomm.h
@@ -19,8 +19,10 @@
 #include <memory>
 
 namespace dazeus {
+namespace db {
+  class Database;
+}
 
-class Database;
 class ConfigReader;
 typedef std::shared_ptr<ConfigReader> ConfigReaderPtr;
 class DaZeus;
@@ -169,7 +171,7 @@ class PluginComm : public NetworkListener
   };
 
   public:
-            PluginComm( Database *d, ConfigReaderPtr c, DaZeus *bot );
+            PluginComm( db::Database *d, ConfigReaderPtr c, DaZeus *bot );
   virtual  ~PluginComm();
   void dispatch(const std::string &event, const std::vector<std::string> &parameters);
   void init();
@@ -191,7 +193,7 @@ class PluginComm : public NetworkListener
     std::vector<int> localServers_;
     std::vector<Command*> commandQueue_;
     std::map<int,SocketInfo> sockets_;
-    Database *database_;
+    db::Database *database_;
     ConfigReaderPtr config_;
     DaZeus *dazeus_;
     std::string handle(int dev, const std::string &line, SocketInfo &info);
